fix: uint8 decimals return type

+++ contracts/LiquidityGauge.vy
@@ -857,7 +857,7 @@ def inflation_rate() -> uint256:
 
 @view
 @external
-def decimals() -> uint256:
+def decimals() -> uint8:
     """
     @notice Get the number of decimals for this token
     @dev Implemented as a view method to reduce gas costs
